Clone original html.Node instances in 2nd pass
The first implementation _just_ set the original html.Node instance as a
field on _out_ Node types. But that means when processing a node during
construction; the html.Node already represents the fully constructed
tree.

To be able to implement insertion steps for certain elements (script
used as the example here), that leads to a wrong result. E.g., the test
created here displays that content _after_ the script-tag, shouldn't be
visible to the script, as it's not in the DOM yet.

That's why all html.Node instances are cloned and reassembled in the
original order.