Refactor global class initialisation
Makes it so; that super class specs don't need to know about inherited
classes. And inherited classes only needs to know the _name_ of the
superclass, not the FunctionTemplate.

In particular, this makes code generation and modularisation more easily
accessible. Gode gen can just create an `init` function registering the
classes it creates, and modules can add new classes with proper
inheritance trees.