Remove ObjectId() from EventTarget
This is part of a solution that makes sure the same JS object is
returned when the same DOM nodes are found, e.g., using
`getElementById`.

Objects that are constructed in JS do not need to use this solution, so
the properties were removed from the interface. Only `Node` as well as
the `Window` object has this solution. The `window` strictly doesn't
need it though.